+++ lib/img/operations.h
@@ -15,7 +15,7 @@ int save_image(ppm_image *img);
 uint8_t* read_pixel_line(ppm_image *img);
 
 // takes a color pixel line of size depth * size_x and converts it to grayscale pixel line
-uint8_t* color_to_gray_line(uint8_t *pixel_line, uint8_t depth, uint8_t size_x);
+uint8_t* color_to_gray_line(uint8_t *pixel_line, uint8_t depth, uintmax_t size_x);
 
 // new_image creates a new (empty) image with the given parameters
 ppm_image *new_image(const char *filename, uintmax_t size_x, uintmax_t size_y, uint8_t depth);
